ML_Backend/requirements.txt
flask==2.3.3
pymongo==4.5.0
pandas==2.2.2
python-dotenv==1.0.0